Move vfs from pkg to model
