+++ model/job/trigger_event.go
@@ -4,12 +4,12 @@ import (
 	"errors"
 	"strings"
 
+	"github.com/cozy/cozy-stack/model/vfs"
 	"github.com/cozy/cozy-stack/pkg/consts"
 	"github.com/cozy/cozy-stack/pkg/couchdb"
 	"github.com/cozy/cozy-stack/pkg/logger"
 	"github.com/cozy/cozy-stack/pkg/permissions"
 	"github.com/cozy/cozy-stack/pkg/realtime"
-	"github.com/cozy/cozy-stack/pkg/vfs"
 )
 
 // EventTrigger implements Trigger for realtime triggered events
