+++ model/vfs/couchdb_indexer.go
