+++ model/vfs/directory.go
