+++ model/vfs/iter.go
