+++ model/vfs/metadata_test.go
