+++ model/vfs/vfsswift/thumbs_v2.go
@@ -8,8 +8,8 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/cozy/cozy-stack/model/vfs"
 	"github.com/cozy/cozy-stack/pkg/prefixer"
-	"github.com/cozy/cozy-stack/pkg/vfs"
 	"github.com/cozy/swift"
 )
